fix windows compile problem
