+++ reader/singlefile_test.go
@@ -8,6 +8,8 @@ import (
 	"strings"
 	"testing"
 
+	"github.com/qiniu/logkit/utils"
+
 	"github.com/stretchr/testify/assert"
 )
 
@@ -33,7 +35,7 @@ func Test_singleFileRotate(t *testing.T) {
 	absPath, err := filepath.Abs(fileName)
 	assert.NoError(t, err)
 	assert.Equal(t, absPath, sf.Source())
-	oldInode := getInode(sf.pfi)
+	oldInode := utils.GetInode(sf.pfi)
 
 	//rotate file(rename old file + create new file)
 	renameTestFile(fileName, fileNameRotated)
@@ -54,7 +56,7 @@ func Test_singleFileRotate(t *testing.T) {
 		t.Error(err)
 	}
 
-	newInode := getInode(sf.pfi)
+	newInode := utils.GetInode(sf.pfi)
 	assert.NotEqual(t, newInode, oldInode)
 
 	assert.Equal(t, 5, n)
@@ -83,7 +85,7 @@ func Test_singleFileNotRotate(t *testing.T) {
 	if err != nil {
 		t.Error(err)
 	}
-	oldInode := getInode(sf.pfi)
+	oldInode := utils.GetInode(sf.pfi)
 
 	//read file 正常读
 	p := make([]byte, 5)
@@ -98,7 +100,7 @@ func Test_singleFileNotRotate(t *testing.T) {
 	n, err = sf.Read(p)
 	assert.Equal(t, io.EOF, err)
 
-	newInode := getInode(sf.pfi)
+	newInode := utils.GetInode(sf.pfi)
 	assert.Equal(t, newInode, oldInode)
 
 	//append文件
